Minimally require Ruby 3.0
Looks like I just bumped to 2.7 in d9b2b01 back when I wrote the gem.

I had accidentally used a Ruby 3.0 syntax feature here
https://github.com/kaspth/active_record-associated_object/blob/9b182b3eca357f0c6620dc655005bea1a62e2db0/lib/active_record/associated_object.rb#L18

And forgot to test on 2.7 in CI, meaning that I never caught that the gem can't load on 2.7.
So we've effectively been 3.0 exclusive forever, so any users on this are already minimally on 3.0 and I'd rather bump to 3.0 since 2.7 is EOL'ed.