Fix: Extensions aren't loaded in development/test
PR #19 previously loaded the associated object class when the has_object
was defined, so that extensions would get run. It seems like that code
got lost in a rebase. So now, extensions only run on boot if eager
loading is enabled (which it is in production). Otherwise they aren't
loaded until the first time the associated object gets used.